# src/BookBarter.React/src/api/generated/models/author-for-moderation-dto-paginated-result.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BookBarter.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { AuthorForModerationDto } from './author-for-moderation-dto';

/**
 * 
 * @export
 * @interface AuthorForModerationDtoPaginatedResult
 */
export interface AuthorForModerationDtoPaginatedResult {
    /**
     * 
     * @type {number}
     * @memberof AuthorForModerationDtoPaginatedResult
     */
    'pageNumber'?: number;
    /**
     * 
     * @type {number}
     * @memberof AuthorForModerationDtoPaginatedResult
     */
    'pageSize'?: number;
    /**
     * 
     * @type {number}
     * @memberof AuthorForModerationDtoPaginatedResult
     */
    'total'?: number;
    /**
     * 
     * @type {Array<AuthorForModerationDto>}
     * @memberof AuthorForModerationDtoPaginatedResult
     */
    'items'?: Array<AuthorForModerationDto> | null;
}