# src/BookBarter.React/src/api/generated/models/author-for-moderation-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BookBarter.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { BookDto } from './book-dto';

/**
 * 
 * @export
 * @interface AuthorForModerationDto
 */
export interface AuthorForModerationDto {
    /**
     * 
     * @type {number}
     * @memberof AuthorForModerationDto
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof AuthorForModerationDto
     */
    'firstName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AuthorForModerationDto
     */
    'middleName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AuthorForModerationDto
     */
    'lastName'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof AuthorForModerationDto
     */
    'approved'?: boolean;
    /**
     * 
     * @type {Array<BookDto>}
     * @memberof AuthorForModerationDto
     */
    'books'?: Array<BookDto> | null;
}