# src/BookBarter.React/src/api/generated/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "BookBarter.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Authors/paged": {
      "post": {
        "tags": [
          "Authors"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetPagedAuthorsQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetPagedAuthorsQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetPagedAuthorsQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorDtoPaginatedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorDtoPaginatedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorDtoPaginatedResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/Authors/paged/moderated": {
      "post": {
        "tags": [
          "Authors"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetPagedAuthorsForModerationQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetPagedAuthorsForModerationQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetPagedAuthorsForModerationQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorForModerationDtoPaginatedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorForModerationDtoPaginatedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorForModerationDtoPaginatedResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/Authors": {
      "post": {
        "tags": [
          "Authors"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuthorCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuthorCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuthorCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Authors/{id}": {
      "put": {
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAuthorCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAuthorCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAuthorCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Authors/{id}/approve": {
      "put": {
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Books/{id}": {
      "get": {
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Books/paged": {
      "post": {
        "tags": [
          "Books"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetPagedBooksQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetPagedBooksQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetPagedBooksQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookDtoPaginatedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDtoPaginatedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDtoPaginatedResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/Books/paged/moderated": {
      "post": {
        "tags": [
          "Books"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetPagedBooksForModerationQuery"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetPagedBooksForModerationQuery"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetPagedBooksForModerationQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookForModerationDtoPaginatedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookForModerationDtoPaginatedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookForModerationDtoPaginatedResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/Books": {
      "post": {
        "tags": [
          "Books"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Books/{id}/approve": {
      "put": {
        "tags": [
          "Books"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AuthorDtoPaginatedResult": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AuthorForModerationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "approved": {
            "type": "boolean"
          },
          "books": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AuthorForModerationDtoPaginatedResult": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorForModerationDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BookDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isbn": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "publicationDate": {
            "type": "string",
            "format": "date"
          },
          "approved": {
            "type": "boolean"
          },
          "genreName": {
            "type": "string",
            "nullable": true
          },
          "publisherName": {
            "type": "string",
            "nullable": true
          },
          "authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BookDtoPaginatedResult": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BookForModerationDto": {
        "required": [
          "addedToDatabaseDate",
          "publicationDate"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isbn": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "publicationDate": {
            "type": "string",
            "format": "date"
          },
          "addedToDatabaseDate": {
            "type": "string",
            "format": "date-time"
          },
          "approved": {
            "type": "boolean"
          },
          "genre": {
            "$ref": "#/components/schemas/GenreDto"
          },
          "publisher": {
            "$ref": "#/components/schemas/PublisherDto"
          },
          "authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorDto"
            },
            "nullable": true
          },
          "ownedByUsers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookRelationshipWithUserDto"
            },
            "nullable": true
          },
          "wantedByUsers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookRelationshipWithUserDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BookForModerationDtoPaginatedResult": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookForModerationDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BookRelationshipWithUserDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/CollapsedUserDto"
          },
          "addedDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "CollapsedUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateAuthorCommand": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateBookCommand": {
        "type": "object",
        "properties": {
          "isbn": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "publicationDate": {
            "type": "string",
            "format": "date"
          },
          "authorsIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "genreId": {
            "type": "integer",
            "format": "int32"
          },
          "publisherId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "GenreDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GetPagedAuthorsForModerationQuery": {
        "type": "object",
        "properties": {
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "orderByProperty": {
            "type": "string",
            "nullable": true
          },
          "orderDirection": {
            "type": "string",
            "nullable": true
          },
          "approved": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "GetPagedAuthorsQuery": {
        "type": "object",
        "properties": {
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "orderByProperty": {
            "type": "string",
            "nullable": true
          },
          "orderDirection": {
            "type": "string",
            "nullable": true
          },
          "query": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetPagedBooksForModerationQuery": {
        "type": "object",
        "properties": {
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "orderByProperty": {
            "type": "string",
            "nullable": true
          },
          "orderDirection": {
            "type": "string",
            "nullable": true
          },
          "approved": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "GetPagedBooksQuery": {
        "type": "object",
        "properties": {
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "orderByProperty": {
            "type": "string",
            "nullable": true
          },
          "orderDirection": {
            "type": "string",
            "nullable": true
          },
          "skipLoggedInUserBooks": {
            "type": "boolean"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "authorName": {
            "type": "string",
            "nullable": true
          },
          "genreId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "publisherId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PublisherDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateAuthorCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateBookCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isbn": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "publicationDate": {
            "type": "string",
            "format": "date"
          },
          "authorsIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "genreId": {
            "type": "integer",
            "format": "int32"
          },
          "publisherId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}